corrected prototypes

+++ include/libroardsp/remove.h
@@ -45,8 +45,8 @@ struct roar_remove_state {
 int roar_remove_init (struct roar_remove_state * state);
 int roar_remove      (void    * inout, void    * subs, int samples, int bits, struct roar_remove_state * state);
 int roar_remove_8    (int8_t  * inout, int8_t  * subs, int samples, struct roar_remove_state * state);
-int roar_remove_16   (int8_t  * inout, int8_t  * subs, int samples, struct roar_remove_state * state);
-int roar_remove_32   (int8_t  * inout, int8_t  * subs, int samples, struct roar_remove_state * state);
+int roar_remove_16   (int16_t * inout, int16_t * subs, int samples, struct roar_remove_state * state);
+int roar_remove_32   (int32_t * inout, int32_t * subs, int samples, struct roar_remove_state * state);
 
 int roar_remove_so   (void    * subout, void    * in, int samples, int bits, struct roar_remove_state * state);
 int roar_remove_so16 (int16_t * subout, int16_t * in, int samples, struct roar_remove_state * state);
