+++ libroardsp/remove.c
@@ -51,7 +51,7 @@ int roar_remove      (void * inout, void * subs, int samples, int bits, struct r
 int roar_remove_8    (int8_t  * inout, int8_t  * subs, int samples, struct roar_remove_state * state) {
  return -1;
 }
-int roar_remove_16   (int8_t  * inout, int8_t  * subs, int samples, struct roar_remove_state * state) {
+int roar_remove_16   (int16_t * inout, int16_t * subs, int samples, struct roar_remove_state * state) {
  int i;
  register int32_t s;
  register int32_t peak;
@@ -83,7 +83,7 @@ int roar_remove_16   (int8_t  * inout, int8_t  * subs, int samples, struct roar_
 
  return 0;
 }
-int roar_remove_32   (int8_t  * inout, int8_t  * subs, int samples, struct roar_remove_state * state) {
+int roar_remove_32   (int32_t * inout, int32_t * subs, int samples, struct roar_remove_state * state) {
  return -1;
 }
